+++ src/main/java/com/lrhealth/data/converge/scheduled/model/dto/ResultViewInfoDto.java
@@ -25,5 +25,5 @@ public class ResultViewInfoDto {
 
     private Integer status;
 
-    private Integer fileSize;
+    private Long fileSize;
 }
